requirements.txt
geoserver-rest